// src/components/Button/Controller/ControllerButton.type.ts
export interface Button
  extends React.HTMLAttributes<HTMLButtonElement | HTMLAnchorElement> {
  children?: React.ReactNode;
  className?: string;
  disabled?: boolean;
  small?: boolean;
  wide?: boolean;
  kind?: 'primary' | 'danger' | 'secondary' | 'ghost' | 'tertiary';
  href?: string;
  tabIndex?: number;
  type?: 'button' | 'reset' | 'submit';
  skeleton?: boolean;
  buttonName?: string;
  buttonDescription?: string;
}